Create README.md
During recent excavations on Mars were discovered sheets of paper with mysterious symbols on them. After much research, scientists have come to the conclusion that the inscriptions on them could actually be ordinary numerical equations. In addition, there was strong evidence from other sources that the Martians knew only three operations - addition, multiplication and subtraction (the Martians never used "unary minus": instead of "-5" they wrote "0-5"). The scientists also proved that the Martians did not give operations different priority, but simply evaluated expressions (if they did not have brackets) from left to right: for example, 3+3*5 they had 30, not 18. Unfortunately, the symbols of arithmetic operations were erased. For example, if there was a record "18=7 (5 3) 2", it is possible to restore this record as "18=7+(5-3)*2". You want to write a program that finds the desired arrangement of characters or reports that it does not exist.